+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed small label that has class `.php8-compare__label` on php 8.3 page
 - Other small fixes and improvements on php 8.3 page
 - Remove 2 old videos from the home page sidebar
+- Add 1 Symfony video to the home page sidebar because I removed one of the old videos for Symfony 4
 
 ----